finished one bot feature but loading time needs improvement
